API/requirements.txt
fastapi
uvicorn
mysql-connector-python
pymysql
sqlalchemy
python-dotenv
